make file structure and refactored codes
