+++ app/schemas/base_response.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class BaseResponse(BaseModel):
+    message: str
