--- app/score_calculator/divide/thirteen_orphans_shape.py
@@ -1,13 +1,0 @@
-from typing import Final
-
-from app.score_calculator.enums.enums import Tile
-from app.score_calculator.hand.hand import Hand
-
-FULLY_HAND_SIZE: Final[int] = 14
-
-
-def can_divide_thirteen_orphans_shape(hand: Hand) -> bool:
-    if len(hand.call_blocks) or sum(hand.tiles) != FULLY_HAND_SIZE:
-        return False
-
-    return all(hand.tiles[tile] >= 1 for tile in Tile.outside_tiles())
