--- app/score_calculator/utility/utility.py
@@ -1,265 +1,0 @@
-from typing import Final
-
-from app.score_calculator.block.block import Block
-from app.score_calculator.enums.enums import Yaku
-
-
-def has_opened_blocks(blocks: list[Block]) -> bool:
-    return any(block.is_opened for block in blocks)
-
-
-YAKUS_INCLUDING_PUNG_OF_TOH: Final[set[Yaku]] = {
-    Yaku.BigFourWinds,
-    Yaku.NineGates,
-    Yaku.AllTerminals,
-    Yaku.AllHonors,
-    Yaku.AllTerminalsAndHonors,
-}
-
-EXCLUDED_YAKUS: Final[dict[Yaku, list[Yaku]]] = {
-    Yaku.BigFourWinds: [
-        Yaku.BigThreeWinds,
-        Yaku.AllPungs,
-        Yaku.PrevalentWind,
-        Yaku.SeatWind,
-        Yaku.PungOfTerminalsOrHonors,
-    ],
-    Yaku.BigThreeDragons: [Yaku.TwoDragonsPungs, Yaku.DragonPung],
-    Yaku.NineGates: [
-        Yaku.FullFlush,
-        Yaku.ConcealedHand,
-        Yaku.PungOfTerminalsOrHonors,
-        Yaku.HalfFlush,
-        Yaku.NoHonorTiles,
-    ],
-    Yaku.FourKongs: [
-        Yaku.SingleWait,
-        Yaku.AllPungs,
-        Yaku.MeldedKong,
-        Yaku.ConcealedKong,
-        Yaku.TwoMeldedKongs,
-        Yaku.TwoConcealedKongs,
-        Yaku.ThreeKongs,
-    ],
-    Yaku.SevenShiftedPairs: [
-        Yaku.FullFlush,
-        Yaku.ConcealedHand,
-        Yaku.SingleWait,
-        Yaku.HalfFlush,
-        Yaku.SevenPairs,
-        Yaku.OneVoidedSuit,
-        Yaku.NoHonorTiles,
-    ],
-    Yaku.AllGreen: [],
-    Yaku.ThirteenOrphans: [
-        Yaku.AllTypes,
-        Yaku.ConcealedHand,
-        Yaku.SingleWait,
-        Yaku.AllTerminalsAndHonors,
-    ],
-    Yaku.AllTerminals: [
-        Yaku.AllPungs,
-        Yaku.OutsideHand,
-        Yaku.PungOfTerminalsOrHonors,
-        Yaku.NoHonorTiles,
-        Yaku.AllTerminalsAndHonors,
-    ],
-    Yaku.LittleFourWinds: [
-        Yaku.BigThreeWinds,
-    ],  # Yaku.PungOfTerminalsOrHonors by used block flag
-    Yaku.LittleThreeDragons: [Yaku.TwoDragonsPungs, Yaku.DragonPung],
-    Yaku.AllHonors: [
-        Yaku.AllPungs,
-        Yaku.OutsideHand,
-        Yaku.PungOfTerminalsOrHonors,
-        Yaku.AllTerminalsAndHonors,
-    ],
-    Yaku.FourConcealedPungs: [
-        Yaku.ThreeConcealedPungs,
-        Yaku.TwoConcealedPungs,
-        Yaku.AllPungs,
-        Yaku.ConcealedHand,
-    ],
-    Yaku.PureTerminalChows: [
-        Yaku.SevenPairs,
-        Yaku.FullFlush,
-        Yaku.AllChows,
-        Yaku.NoHonorTiles,
-        Yaku.PureDoubleChow,
-        Yaku.TwoTerminalChows,
-        Yaku.HalfFlush,
-        Yaku.OneVoidedSuit,
-    ],
-    Yaku.QuadrupleChow: [Yaku.PureShiftedPungs, Yaku.TileHog, Yaku.PureDoubleChow],
-    Yaku.FourPureShiftedPungs: [Yaku.PureTripleChow, Yaku.AllPungs],
-    Yaku.FourPureShiftedChows: [Yaku.ShortStraight, Yaku.TwoTerminalChows],
-    Yaku.ThreeKongs: [
-        Yaku.MeldedKong,
-        Yaku.ConcealedKong,
-        Yaku.TwoMeldedKongs,
-        Yaku.TwoConcealedKongs,
-    ],
-    Yaku.AllTerminalsAndHonors: [
-        Yaku.AllPungs,
-        Yaku.OutsideHand,
-        Yaku.PungOfTerminalsOrHonors,
-    ],
-    Yaku.SevenPairs: [Yaku.ConcealedHand, Yaku.SingleWait],
-    Yaku.GreaterHonorsAndKnittedTiles: [
-        Yaku.AllTypes,
-        Yaku.ConcealedHand,
-        Yaku.LesserHonorsAndKnittedTiles,
-    ],
-    Yaku.AllEvenPungs: [Yaku.AllPungs, Yaku.AllSimples, Yaku.NoHonorTiles],
-    Yaku.FullFlush: [Yaku.NoHonorTiles, Yaku.OneVoidedSuit, Yaku.HalfFlush],
-    Yaku.PureTripleChow: [Yaku.PureShiftedPungs, Yaku.PureDoubleChow],
-    Yaku.PureShiftedPungs: [Yaku.PureTripleChow],
-    Yaku.UpperTiles: [Yaku.NoHonorTiles, Yaku.UpperFour],
-    Yaku.MiddleTiles: [Yaku.NoHonorTiles, Yaku.AllSimples],
-    Yaku.LowerTiles: [Yaku.NoHonorTiles, Yaku.LowerFour],
-    Yaku.PureStraight: [],
-    Yaku.ThreeSuitedTerminalChows: [
-        Yaku.AllChows,
-        Yaku.NoHonorTiles,
-        Yaku.MixedDoubleChow,
-        Yaku.TwoTerminalChows,
-    ],
-    Yaku.PureShiftedChows: [],
-    Yaku.AllFives: [Yaku.AllSimples, Yaku.NoHonorTiles],
-    Yaku.TriplePung: [Yaku.DoublePung],
-    Yaku.ThreeConcealedPungs: [Yaku.TwoConcealedPungs],
-    Yaku.LesserHonorsAndKnittedTiles: [Yaku.AllTypes, Yaku.ConcealedHand],
-    Yaku.KnittedStraight: [],
-    Yaku.UpperFour: [Yaku.NoHonorTiles],
-    Yaku.LowerFour: [Yaku.NoHonorTiles],
-    Yaku.BigThreeWinds: [],  # Yaku.PungOfTerminalsOrHonors by used block flag
-    Yaku.MixedStraight: [],
-    Yaku.ReversibleTiles: [Yaku.OneVoidedSuit],
-    Yaku.MixedTripleChow: [],
-    Yaku.MixedShiftedPungs: [],
-    Yaku.ChickenHand: [],
-    Yaku.LastTileDraw: [Yaku.SelfDrawn],
-    Yaku.LastTileClaim: [],
-    Yaku.OutWithReplacementTile: [Yaku.SelfDrawn],
-    Yaku.RobbingTheKong: [Yaku.LastTile],
-    Yaku.TwoConcealedKongs: [
-        Yaku.TwoMeldedKongs,
-        Yaku.ConcealedKong,
-        Yaku.TwoConcealedPungs,
-    ],
-    Yaku.AllPungs: [],
-    Yaku.HalfFlush: [Yaku.OneVoidedSuit],
-    Yaku.MixedShiftedChows: [],
-    Yaku.AllTypes: [],
-    Yaku.MeldedHand: [Yaku.SingleWait],
-    Yaku.TwoDragonsPungs: [Yaku.DragonPung],
-    Yaku.OutsideHand: [],
-    Yaku.FullyConcealedHand: [Yaku.SelfDrawn],
-    Yaku.TwoMeldedKongs: [Yaku.MeldedKong],
-    Yaku.LastTile: [],
-    Yaku.DragonPung: [],
-    Yaku.PrevalentWind: [],
-    Yaku.SeatWind: [],
-    Yaku.ConcealedHand: [],
-    Yaku.AllChows: [Yaku.NoHonorTiles],
-    Yaku.TileHog: [],  # if (AllGreen or AllTerminals) and SevenPairs: must be removed
-    Yaku.DoublePung: [],
-    Yaku.TwoConcealedPungs: [],
-    Yaku.ConcealedKong: [],
-    Yaku.AllSimples: [Yaku.NoHonorTiles],
-    Yaku.PureDoubleChow: [],
-    Yaku.MixedDoubleChow: [],
-    Yaku.ShortStraight: [],
-    Yaku.TwoTerminalChows: [],
-    Yaku.PungOfTerminalsOrHonors: [],
-    Yaku.MeldedKong: [],
-    Yaku.OneVoidedSuit: [],
-    Yaku.NoHonorTiles: [],
-    Yaku.EdgeWait: [],
-    Yaku.ClosedWait: [],
-    Yaku.SingleWait: [],
-    Yaku.SelfDrawn: [],
-}
-
-YAKU_POINT: Final[dict[Yaku, int]] = {
-    Yaku.ChickenHand: 8,
-    Yaku.SevenShiftedPairs: 88,
-    Yaku.ThirteenOrphans: 88,
-    Yaku.BigFourWinds: 88,
-    Yaku.BigThreeDragons: 88,
-    Yaku.NineGates: 88,
-    Yaku.AllGreen: 88,
-    Yaku.FourKongs: 88,
-    Yaku.FourConcealedPungs: 64,
-    Yaku.AllTerminals: 64,
-    Yaku.LittleFourWinds: 64,
-    Yaku.LittleThreeDragons: 64,
-    Yaku.AllHonors: 64,
-    Yaku.PureTerminalChows: 64,
-    Yaku.QuadrupleChow: 48,
-    Yaku.FourPureShiftedPungs: 48,
-    Yaku.FourPureShiftedChows: 32,
-    Yaku.ThreeKongs: 32,
-    Yaku.AllTerminalsAndHonors: 32,
-    Yaku.SevenPairs: 24,
-    Yaku.GreaterHonorsAndKnittedTiles: 24,
-    Yaku.AllEvenPungs: 24,
-    Yaku.FullFlush: 24,
-    Yaku.UpperTiles: 24,
-    Yaku.MiddleTiles: 24,
-    Yaku.LowerTiles: 24,
-    Yaku.PureTripleChow: 24,
-    Yaku.PureShiftedPungs: 24,
-    Yaku.PureStraight: 16,
-    Yaku.ThreeSuitedTerminalChows: 16,
-    Yaku.PureShiftedChows: 16,
-    Yaku.AllFives: 16,
-    Yaku.TriplePung: 16,
-    Yaku.ThreeConcealedPungs: 16,
-    Yaku.LesserHonorsAndKnittedTiles: 12,
-    Yaku.KnittedStraight: 12,
-    Yaku.UpperFour: 12,
-    Yaku.LowerFour: 12,
-    Yaku.BigThreeWinds: 12,
-    Yaku.LastTileDraw: 8,
-    Yaku.LastTileClaim: 8,
-    Yaku.OutWithReplacementTile: 8,
-    Yaku.RobbingTheKong: 8,
-    Yaku.MixedStraight: 8,
-    Yaku.MixedTripleChow: 8,
-    Yaku.ReversibleTiles: 8,
-    Yaku.MixedShiftedPungs: 8,
-    Yaku.TwoConcealedKongs: 8,
-    Yaku.MeldedHand: 6,
-    Yaku.MixedShiftedChows: 6,
-    Yaku.AllPungs: 6,
-    Yaku.HalfFlush: 6,
-    Yaku.AllTypes: 6,
-    Yaku.TwoDragonsPungs: 6,
-    Yaku.FullyConcealedHand: 4,
-    Yaku.LastTile: 4,
-    Yaku.OutsideHand: 4,
-    Yaku.TwoMeldedKongs: 4,
-    Yaku.ConcealedHand: 2,
-    Yaku.DragonPung: 2,
-    Yaku.PrevalentWind: 2,
-    Yaku.SeatWind: 2,
-    Yaku.AllChows: 2,
-    Yaku.DoublePung: 2,
-    Yaku.TwoConcealedPungs: 2,
-    Yaku.TileHog: 2,
-    Yaku.ConcealedKong: 2,
-    Yaku.AllSimples: 2,
-    Yaku.PureDoubleChow: 1,
-    Yaku.MixedDoubleChow: 1,
-    Yaku.ShortStraight: 1,
-    Yaku.TwoTerminalChows: 1,
-    Yaku.PungOfTerminalsOrHonors: 1,
-    Yaku.OneVoidedSuit: 1,
-    Yaku.NoHonorTiles: 1,
-    Yaku.MeldedKong: 1,
-    Yaku.EdgeWait: 1,
-    Yaku.SingleWait: 1,
-    Yaku.ClosedWait: 1,
-    Yaku.SelfDrawn: 1,
-}
