+++ app/services/score_calculator/enums/enums.py
