+++ app/services/score_calculator/winning_conditions/winning_conditions.py
@@ -1,0 +1,16 @@
+from dataclasses import dataclass
+
+from app.services.score_calculator.enums.enums import Tile, Wind
+
+
+@dataclass
+class WinningConditions:
+    winning_tile: Tile
+    is_discarded: bool
+    is_last_tile_in_the_game: bool
+    is_last_tile_of_its_kind: bool
+    is_replacement_tile: bool
+    is_robbing_the_kong: bool
+    count_tenpai_tiles: int
+    seat_wind: Wind
+    round_wind: Wind
